+++ dparse/__init__.py
@@ -1,9 +1,12 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, absolute_import
-"""Top-level package for Dependency Parser."""
 
-__author__ = """Jannis Gebauer"""
-__email__ = 'support@pyup.io'
-__version__ = '0.5.2a'
+# SPDX-License-Identifier: MIT
+# Copyright (c)  Jannis Gebauer and others
+# Originally from https://github.com/pyupio/dparse/
+# Now maintained at https://github.com/nexB/dparse2
+
+# Originally from 
+# __author__ = """Jannis Gebauer"""
+# __email__ = 'support@pyup.io'
+# __version__ = '0.5.2a'
 
 from .parser import parse  # noqa