+++ dparse/errors.py
@@ -1,5 +1,10 @@
 # -*- coding: utf-8 -*-
-from __future__ import unicode_literals, absolute_import
+
+# SPDX-License-Identifier: MIT
+# Copyright (c)  Jannis Gebauer and others
+# Originally from https://github.com/pyupio/dparse/
+# Now maintained at https://github.com/nexB/dparse2
+
 
 class UnknownDependencyFileError(Exception):
     """